Applied various patches to fix the following warnings/assertions:?\011- wxString::formt assert - TabGroup code was not compiling with < wx2.8.8 - Recently opended workspace fixed to work with wx2.9 All fixed were provided by DavidGH
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@4440 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ sdk/codelite_indexer/network/clindexerprotocol.cpp
@@ -36,9 +36,9 @@ bool clIndexerProtocol::ReadReply(clNamedPipe* conn, clIndexerReply& reply)
 	}
 
 	if (actual_read != sizeof(buff_len)) {
-		fprintf(stderr, "ERROR: ReadReply: Protocol error: expected %u bytes, got %u. reason: %d\n",
+		fprintf(stderr, "ERROR: ReadReply: Protocol error: expected %lu bytes, got %u. reason: %d\n",
 				sizeof(buff_len),
-				actual_read,
+				(unsigned int)actual_read,
 				conn->getLastError());
 		return false;
 	}
@@ -55,7 +55,7 @@ bool clIndexerProtocol::ReadReply(clNamedPipe* conn, clIndexerReply& reply)
 	size_t bytes_read(0);
 	while (bytes_left > 0) {
 		if ( !conn->read(data+bytes_read, bytes_left, &actual_read, 10000) ) {
-			fprintf(stderr, "ERROR: Protocol error: expected %u bytes, got %u\n", buff_len, actual_read);
+			fprintf(stderr, "ERROR: Protocol error: expected %u bytes, got %u\n", (unsigned int)buff_len, (unsigned int)actual_read);
 			return false;
 		}
 		bytes_left -= actual_read;
@@ -78,7 +78,7 @@ bool clIndexerProtocol::ReadRequest(clNamedPipe* conn, clIndexerRequest& req)
 	}
 
 	if (actual_read != sizeof(buff_len)) {
-		fprintf(stderr, "ERROR: Protocol error: expected %u bytes, got %u\n", sizeof(buff_len), actual_read);
+		fprintf(stderr, "ERROR: Protocol error: expected %lu bytes, got %u\n", sizeof(buff_len), (unsigned int)actual_read);
 		return false;
 	}
 
@@ -92,7 +92,7 @@ bool clIndexerProtocol::ReadRequest(clNamedPipe* conn, clIndexerRequest& req)
 	size_t bytes_read(0);
 	while (bytes_left > 0) {
 		if ( !conn->read(data+bytes_read, bytes_left, &actual_read, -1) ) {
-			fprintf(stderr, "ERROR: [%s] Protocol error: expected %u bytes, got %u\n", __PRETTY_FUNCTION__, buff_len, actual_read);
+			fprintf(stderr, "ERROR: [%s] Protocol error: expected %u bytes, got %u\n", __PRETTY_FUNCTION__, (unsigned int)buff_len, (unsigned int)actual_read);
 			return false;
 		}
 		bytes_left -= actual_read;
